Made getUser and putUser in service layer
Implemented getUser and putUser as much as possible without more details. I will need to know how we are handling PUT request for user. Are we going to update the entire record? (I am in favor of that because we don't need to check for each specific change (we still validate data though) so we can just update the whole row. This makes it so we don't need different functions for changing every possible attribute in user table. And the component that displays the edit user profile page can just send the entire form data and anything that is changed will naturally overwrite existing data and anything that was the same will stay the same. This is my opinion so please tell me your thoughts.)